+++ python/GafferImageUI/CatalogueUI.py
@@ -38,6 +38,8 @@
 import functools
 import imath
 
+from collections import namedtuple, OrderedDict
+
 import IECore
 
 import Gaffer
@@ -50,6 +52,69 @@ import GafferImageUI
 ## the scope of changing this is somewhat large. When we need to do folders,
 ## then this will probably force us to sort this out. For now, is at least
 ## contained within just this file...
+
+##########################################################################
+# Column Configuration
+#
+# All of the Catalogue's columns are configurable. Columns are identified by
+# a unique name along with a display title, value provider and type. Gaffer
+# provides a default set of columns that can be extended or replaced via
+# the registration of new names or re-registration of existing ones.
+#
+# The columns displayed by a Catalogue are controlled by its
+# "catalogue:columns" metadata value [StringVectorData].
+#
+# Value providers for Standard type columns must return basic-typed IECore.Data
+# (incl. DateTimeData), those for Icon type columns return the name of an image
+# on Gaffer's image paths. They are called with two arguments:
+#
+#  - image : A GafferImage.Catalogue.Image plug (not to be confused with
+#      a standard ImagePlug).
+#
+#  - catalogue: The GafferImage.Catalogue instance attached to the UI.
+#
+#  A suitable context is scoped around the call such that catalogue["out"] will
+#  provide the image for the row the value is being generated for, rather than
+#  the user's current selection.
+##########################################################################
+
+__registeredColumns = OrderedDict()
+
+ColumnType = IECore.Enum.create( "Standard", "Icon" )
+ColumnDefinition = namedtuple( "ColumnDefinition", ( "title", "valueProvider", "type" ) )
+
+# We don't take a ColumnDefinition here as their constructors are a little
+# opaque, and this makes it a bit easier to work out the right order.
+def registerColumn( name, title, valueProvider, type = ColumnType.Standard ) :
+
+	definition = ColumnDefinition( title, valueProvider, type )
+	__registeredColumns[ name ] = definition
+
+def deregisterColumn( name ) :
+
+	if name in __registeredColumns :
+		del __registeredColumns[ name ]
+
+def columnDefinition( name ) :
+
+	return __registeredColumns.get( name, None )
+
+def registeredColumns() :
+
+	return __registeredColumns.keys()
+
+# Standard Columns
+
+registerColumn( "name", "Name", lambda image, _ : image.getName() )
+
+Gaffer.Metadata.registerValue(
+	GafferImage.Catalogue, "catalogue:columns",
+	IECore.StringVectorData( [ "name" ] )
+)
+
+##########################################################################
+# Node registration
+##########################################################################
 
 Gaffer.Metadata.registerNode(
 
@@ -158,7 +223,7 @@ def __viewerKeyPress( viewer, event ) :
 	if not isinstance( viewer.view(), GafferImageUI.ImageView ) :
 		return False
 
-	catalogue = __findCatalogue( viewer.view() )
+	catalogue = __findUpstreamCatalogue( viewer.view() )
 	if catalogue is None :
 		return False
 
@@ -166,7 +231,7 @@ def __viewerKeyPress( viewer, event ) :
 
 	return True
 
-def __findCatalogue( node ) :
+def __findUpstreamCatalogue( node ) :
 
 	catalogue = node.ancestor( GafferImage.Catalogue )
 	if catalogue is not None :
@@ -176,7 +241,7 @@ def __findCatalogue( node ) :
 			upstreamPlug = inPlug.source()
 			if upstreamPlug == inPlug or upstreamPlug.node() == node :
 				continue
-			catalogue = __findCatalogue( upstreamPlug.node() )
+			catalogue = __findUpstreamCatalogue( upstreamPlug.node() )
 			if catalogue is not None :
 				return catalogue
 
@@ -219,6 +284,22 @@ def __incrementImageIndex( catalogue, direction ) :
 # _CataloguePath
 ##########################################################################
 
+def _findSourceCatalogue( imagesPlug ) :
+
+	def walk( plug ) :
+
+		if isinstance( plug.parent(), GafferImage.Catalogue ) :
+			return plug.parent()
+
+		for output in plug.outputs() :
+			r = walk( output )
+			if r is not None :
+				return r
+
+		return None
+
+	return walk( imagesPlug )
+
 class _ImagesPath( Gaffer.Path ) :
 
 	indexMetadataName = 'image:index'
@@ -228,6 +309,7 @@ class _ImagesPath( Gaffer.Path ) :
 		Gaffer.Path.__init__( self, path, root, filter )
 
 		self.__images = images
+		self.__catalogue = _findSourceCatalogue( images )
 
 	def copy( self ) :
 
@@ -236,6 +318,28 @@ class _ImagesPath( Gaffer.Path ) :
 	def isLeaf( self ) :
 
 		return len( self ) > 0
+
+	def propertyNames( self ) :
+
+		return Gaffer.Path.propertyNames( self ) + registeredColumns()
+
+	def property( self, name ) :
+
+		if name not in registeredColumns() :
+			return Gaffer.Path.property( self, name )
+
+		definition = columnDefinition( name )
+
+		imageName = self[ -1 ]
+		image = self.__images[ imageName ]
+
+		# The Catalog API supports overriding the active image
+		# via a context variable, this allows the value provider
+		# to just use catalog["out"] to get to the correct image
+		# without needing to understand the internal workings.
+		with Gaffer.Context(  self.__catalogue.scriptNode().context() ) as context :
+			context[ "catalogue:imageName" ] = imageName
+			return definition.valueProvider( image, self.__catalogue )
 
 	def _orderedImages( self ) :
 
@@ -309,14 +413,16 @@ class _ImageListing( GafferUI.PlugValueWidget ) :
 
 		with self.__column :
 
+			columns = self.__listingColumns()
+
 			self.__pathListing = GafferUI.PathListingWidget(
 				_ImagesPath( self.__images(), [] ),
-				columns = ( GafferUI.PathListingWidget.defaultNameColumn, ),
+				columns = columns,
 				allowMultipleSelection = True
 			)
 			self.__pathListing.setDragPointer( "" )
 			self.__pathListing.setSortable( False )
-			self.__pathListing.setHeaderVisible( False )
+			self.__pathListing.setHeaderVisible( len(columns) >  1 )
 			self.__pathListing.selectionChangedSignal().connect(
 				Gaffer.WeakMethod( self.__pathListingSelectionChanged ), scoped = False
 			)
@@ -391,21 +497,33 @@ class _ImageListing( GafferUI.PlugValueWidget ) :
 
 		self.__column.setEnabled( self._editable() )
 
+	def __listingColumns( self ) :
+
+		columns = []
+
+		for name in Gaffer.Metadata.value( self.__catalogue(), "catalogue:columns" ) :
+
+			definition = columnDefinition( name )
+
+			if not definition :
+				IECore.msg(
+					IECore.Msg.Level.Error,
+					"GafferImageUI.CatalogueUI", "No column registered with name '%s'" % name
+				)
+				continue
+
+			if definition.type == ColumnType.Icon :
+				column = GafferUI.PathListingWidget.IconColumn( definition.title, "", name )
+			else :
+				column = GafferUI.PathListingWidget.StandardColumn( definition.title, name )
+
+			columns.append( column )
+
+		return columns
+
 	def __catalogue( self ) :
 
-		def walk( plug ) :
-
-			if isinstance( plug.parent(), GafferImage.Catalogue ) :
-				return plug.parent()
-
-			for output in plug.outputs() :
-				r = walk( output )
-				if r is not None :
-					return r
-
-			return None
-
-		return walk( self.getPlug() )
+		return _findSourceCatalogue( self.getPlug() )
 
 	def __images( self ) :